share common elasticsearch docker base into common-lib new testlib package
