+++ media-api/test/lib/elasticsearch/ElasticSearchTestBase.scala
@@ -1,61 +1,22 @@
 package lib.elasticsearch
 
-import java.util.UUID
 import com.gu.mediaservice.model._
+import com.gu.mediaservice.testlib.ElasticSearchDockerBase
 import org.joda.time.DateTime
 import org.scalatest.concurrent.PatienceConfiguration.{Interval, Timeout}
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.funspec.AnyFunSpec
-import org.scalatest.time.{Milliseconds, Seconds, Span}
-import org.scalatest.BeforeAndAfterAll
 import org.scalatest.matchers.should.Matchers
-import org.testcontainers.containers.wait.strategy.Wait
-import org.testcontainers.elasticsearch.ElasticsearchContainer
+import org.scalatest.time.{Milliseconds, Seconds, Span}
 import play.api.libs.json.JsString
 
-import java.util.concurrent.atomic.AtomicReference
-import scala.util.Properties
-import scala.concurrent.duration._
-import scala.compat.java8.DurationConverters._
-import scala.jdk.CollectionConverters._
+import java.util.UUID
 
 
-trait ElasticSearchTestBase extends AnyFunSpec with BeforeAndAfterAll with Matchers with ScalaFutures with Fixtures with ConditionFixtures {
+trait ElasticSearchTestBase extends AnyFunSpec with ElasticSearchDockerBase with Matchers with ScalaFutures with Fixtures with ConditionFixtures {
 
   val interval = Interval(Span(100, Milliseconds))
   val timeout = Timeout(Span(10, Seconds))
-
-  val useEsDocker = Properties.envOrElse("USE_DOCKER_FOR_TESTS", "true").toBoolean
-
-  val esContainer: Option[ElasticsearchContainer] = if (useEsDocker) {
-    {
-      val container = new ElasticsearchContainer("docker.elastic.co/elasticsearch/elasticsearch:7.16.2")
-        .withExposedPorts(9200)
-        .withAccessToHost(true)
-        .withEnv(Map(
-          "cluster.name" -> "media-service",
-          "xpack.security.enabled" -> "false",
-          "discovery.type" -> "single-node",
-          "network.host" -> "0.0.0.0"
-        ).asJava)
-        .waitingFor(Wait.forHttp("/")
-          .forPort(9200)
-          .forStatusCode(200)
-          .withStartupTimeout(180.seconds.toJava)
-        )
-      container.start()
-      Some(container)
-    }
-  } else None
-
-  val esPort = esContainer.map(_.getMappedPort(9200)).getOrElse(9200)
-  val es6TestUrl = Properties.envOrElse("ES6_TEST_URL", s"http://localhost:$esPort")
-
-  override protected def afterAll(): Unit = {
-    super.afterAll()
-
-    esContainer foreach { _.stop() }
-  }
 
   lazy val images = Seq(
     createImage("getty-image-1", Agency("Getty Images")),
